Imitate CredHub to keep secrets out of `cf env`
This commit adds a webserver which acts like CredHub does in Pivotal's
CredHub Service Broker. When called by an App Lifecycle, the app
identity is validated over Mutual TLS using CF Instance Identity
certificates [1].

Essentially this is used so that service brokers can put non-sensitive
data in `VCAP_SERVICES` and then get it transformed into credentials
when the app instances start. This keeps secrets out of
`/v3/apps/:GUID` and `cf env`. That is important because those
endpoints have many constant innocuous uses and so their usage can't
be audited effectively.

More info on this whole setup is available in [2].

[1] https://docs.cloudfoundry.org/devguide/deploy-apps/instance-identity.html
[2] https://hackmd.cloudapps.digital/q-GMrPU-Tz29HjlZSwf0Jg